Redo the timing implementation
The lockstep method implemented doesn't work. In fact, real lockstep
will be very difficult to achieve since neither the simulation nor the
controller know the publication rates of each other. Thus, the code
has to work on best effort in current scenario. On the other hand,
best effort situation also represents the reality better. If the flight
controller gets stuck for some reason then it is a catastrphy in real
life as well.